Put GameProvider in useGame file for consistency
